+++ components/pipedrive/src/actions/dealFields/index.ts
@@ -1,4 +1,14 @@
 import { deleteDealField } from "./deleteDealField";
+import { deleteDealFieldV2 } from "./deleteDealFieldV2";
 import { getDealField } from "./getDealField";
+import { getDealFieldV2 } from "./getDealFieldV2";
 import { getDealFields } from "./getDealFields";
-export default { getDealFields, getDealField, deleteDealField };
+import { getDealFieldsV2 } from "./getDealFieldsV2";
+export default {
+  getDealFields,
+  getDealFieldsV2,
+  getDealField,
+  getDealFieldV2,
+  deleteDealField,
+  deleteDealFieldV2,
+};