+++ components/pipedrive/src/actions/personFields/index.ts
@@ -1,4 +1,14 @@
 import { deletePersonField } from "./deletePersonField";
+import { deletePersonFieldV2 } from "./deletePersonFieldV2";
 import { getPersonField } from "./getPersonField";
+import { getPersonFieldV2 } from "./getPersonFieldV2";
 import { getPersonFields } from "./getPersonFields";
-export default { getPersonFields, getPersonField, deletePersonField };
+import { getPersonFieldsV2 } from "./getPersonFieldsV2";
+export default {
+  getPersonFields,
+  getPersonFieldsV2,
+  getPersonField,
+  getPersonFieldV2,
+  deletePersonField,
+  deletePersonFieldV2,
+};