+++ components/pipedrive/src/actions/productFields/index.ts
@@ -1,4 +1,14 @@
 import { deleteProductField } from "./deleteProductField";
+import { deleteProductFieldV2 } from "./deleteProductFieldV2";
 import { getProductField } from "./getProductField";
+import { getProductFieldV2 } from "./getProductFieldV2";
 import { getProductFields } from "./getProductFields";
-export default { getProductFields, deleteProductField, getProductField };
+import { getProductFieldsV2 } from "./getProductFieldsV2";
+export default {
+  getProductFields,
+  getProductFieldsV2,
+  getProductField,
+  getProductFieldV2,
+  deleteProductField,
+  deleteProductFieldV2,
+};