+++ components/pipedrive/src/util/index.ts
@@ -218,6 +218,8 @@ export const getBaseUrl = (version: string) => {
 };
 export const formatPipedriveTimestamp = (date: Date): string =>
   date.toISOString().replace("T", " ").substring(0, 19);
+export const formatPipedriveTimestampV2 = (date: Date): string =>
+  date.toISOString();
 export const fetchRecentsSince = async (
   client: HttpClient,
   sinceTimestamp: string,